pip_requirements.txt
noise
pip
scipy
numpy
pillow
configargparse
webcolors